+++ src/main/java/likelion/scent_yonsei/domain/makers/api/dto/LikedRes.java
@@ -1,0 +1,4 @@
+package likelion.scent_yonsei.domain.makers.api.dto;
+
+public record LikedRes(Long count) {
+}
