--- src/main/java/likelion/scent_yonsei/makers/api/dto/LikedReq.java
@@ -1,4 +1,0 @@
-package likelion.scent_yonsei.makers.api.dto;
-
-public record LikedReq(int count) {
-}
